test: split `index.test.ts` into per-module suites
